requirements.txt
# Core dependencies
fastapi==0.104.1
websockets==12.0
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database and storage
supabase>=2.0.3  # Allow newer versions for httpx compatibility
redis==5.0.1
asyncpg==0.29.0

# AI for semantic routing
google-generativeai==0.3.2
# pydantic-ai>=0.7.0  # Temporarily disabled due to httpx conflict with supabase

# Diagram generation
matplotlib==3.8.2
plotly==5.18.0
kaleido==0.2.1  # For plotly image export
pillow==10.1.0

# Utilities
python-dotenv==1.0.0
logfire==0.20.0
tiktoken==0.5.2  # For token counting

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
mypy==1.7.1